Create My session and User manager
My session is easier and faster than the original one.
Instead of Map I use an object.

The user manager detects new users and users in a given session.
It also waits until the user id is defined and then executes all commands.
This is needed when the user sends messages when the bot is off.