kata.js
kata.js is a module for working with katas more conveniently.
It has almost all the functions for working with kata.
Work with both the database and dropbox storage at the same time.

К другим изменениям:

Auto-renew access token in dropbox.js.
Additional functions are now available not only for client but also for PG in pg.js.
A few new features in toin.js